Fix: placed public accesable filament pages outside of panel auto discover
