+++ tests/Feature/OrganisationInviteTest.php
@@ -3,7 +3,7 @@
 use App\Enums\OrganisationRole;
 use App\Enums\Role;
 use App\Filament\Organiser\Clusters\Settings\Resources\UserResource\Pages\ListOrganiserUsers;
-use App\Filament\Organiser\Pages\AcceptOrganisationInvite;
+use App\Filament\Public\Pages\AcceptOrganisationInvite;
 use App\Mail\OrganisationInviteMail;
 use App\Models\Organisation;
 use App\Models\OrganisationInvite;
